simclr/requirements.txt
# PyTorch and related packages
torch>=2.0.0
torchvision>=0.15.0
tensorboard>=2.14.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Machine learning utilities
scikit-learn>=1.0.0

# Image processing
Pillow>=8.0.0

# Progress bars and utilities
tqdm>=4.64.0

# Plotting (optional, for visualization)
matplotlib>=3.5.0
seaborn>=0.11.0

# Development tools (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
